refactor: registry에 add 메서드 추가
